rework send_fake: copy as much as possible from the original event
send_fake now copies almost all field from the original event, except
for the the window, keycode and state field.

When we grab a key we grab it onto the root window, so every event we
receive has a field window that is equal to the root.  (this will
change in the future), so we need to change that.

The keycode and state obviously need to be changed, it's the whole
point of the function.

It doesn't seem to fix the known bug that occur in dino thought.