grabkey only on the targeted windows
now we grabkey() only on the window that matches what the user
requested.  That is, given

        match class 'Firefox'
        on "C-n" do "<Down>"

we only bind C-n on windows that match the class Firefox, and not on
the root window as we did before.

In order to do this, we traverse the tree during startup and grabkey
on matching windows, than we subscribe also to SubstructureNotify so
that when a window gets mapped we receive an event.

There are still some downsides to this.  Firefox uses also
'/usr/local/lib/firefox/firefox' as class name on my machine, and
sometimes it seems like we don’t grab the key at the first try (maybe
Firefox changes the class after the window gets mapped?)